Updated payment Page to add date of end of Service

+++ WebApp/expressPay.php
@@ -8,16 +8,16 @@ $chosenService = $q->fetchAll();
 $q = $DbManager->getDb()->prepare("SELECT idOrders from Orders WHERE status = 'express'");
 $q->execute();
 $res = $q->fetchAll();
-
-$q = $DbManager->getDb()->prepare("DELETE FROM Orders WHERE idOrders = ?");
-$q->execute([
-   $res[0]['idOrders']
-]);
-$q = $DbManager->getDb()->prepare("DELETE FROM orderOptions WHERE idOrders = ?");
-$q->execute([
-    $res[0]['idOrders']
-]);
-
+if(!empty($res)) {
+    $q = $DbManager->getDb()->prepare("DELETE FROM Orders WHERE idOrders = ?");
+    $q->execute([
+        $res[0]['idOrders']
+    ]);
+    $q = $DbManager->getDb()->prepare("DELETE FROM orderOptions WHERE idOrders = ?");
+    $q->execute([
+        $res[0]['idOrders']
+    ]);
+}
 
 function getDates($duration, $option)
 {
@@ -56,6 +56,13 @@ switch ($_GET['package']){
             [
             DbManager::v4(),
             $uuid,
+            "timeVariable",
+            $_GET['hour']
+        ]);
+        $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
+            [
+            DbManager::v4(),
+            $uuid,
             "date",
             $_GET['date']
         ]);
@@ -75,7 +82,7 @@ switch ($_GET['package']){
         }else{
             $price = $_GET['day'] * $_GET['hour'] * $chosenService[0]['price'] * 1/8;
         }
-        if($week1 === "on") {
+        if($week1 == "on") {
             if($_GET['day'] <= 7) {
                 ++$options;
             }
@@ -87,7 +94,7 @@ switch ($_GET['package']){
                 $week1
             ]);
         }
-        if($week2 === "on") {
+        if($week2 == "on") {
             if($_GET['day'] <= 7) {
                 ++$options;
             }
@@ -99,7 +106,7 @@ switch ($_GET['package']){
                 $week2
             ]);
         }
-        if ($week3 === "on") {
+        if ($week3 == "on") {
             if($_GET['day'] <= 7) {
                 ++$options;
             }
@@ -111,7 +118,7 @@ switch ($_GET['package']){
                 $week3
             ]);
         }
-        if($week4 === "on") {
+        if($week4 == "on") {
             if($_GET['day'] <= 7) {
                 ++$options;
             }
@@ -123,7 +130,7 @@ switch ($_GET['package']){
                 $week4
             ]);
         }
-        if($week5 === "on") {
+        if($week5 == "on") {
             if($_GET['day'] <= 7) {
                 ++$options;
             }
@@ -135,7 +142,7 @@ switch ($_GET['package']){
                 $week5
             ]);
         }
-        if($week6 === "on") {
+        if($week6 == "on") {
             if($_GET['day'] <= 7) {
                 $options += 1;
             }
@@ -147,7 +154,7 @@ switch ($_GET['package']){
                 $week6
             ]);
         }
-        if($week7 === "on") {
+        if($week7 == "on") {
             if($_GET['day'] <= 7) {
                 ++$options;
             }
@@ -159,6 +166,15 @@ switch ($_GET['package']){
                 $week7
             ]);
         }
+        $endService = date('Y-m-d', time() + $_GET['day'] * 24 * 60 * 60);
+
+        $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
+            [
+                DbManager::v4(),
+                $uuid,
+                "timeVariable",
+                $_GET['day']
+            ]);
         if($_GET['day'] <= 7) {
             $price *= $options;
         }
@@ -187,7 +203,7 @@ switch ($_GET['package']){
         $price = $_GET['hour'] * $chosenService[0]['price'] * 1 / 8;
         $options = 0;
 
-        if($month1 === "on") {
+        if($month1 == "on") {
             $options += getDates(31*$_GET['day'], 1);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -197,7 +213,7 @@ switch ($_GET['package']){
                 $month1
             ]);
         }
-        if($month2 === "on") {
+        if($month2 == "on") {
             $options += getDates(31*$_GET['day'], 2);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -207,7 +223,7 @@ switch ($_GET['package']){
                 $month2
             ]);
         }
-        if($month3 === "on") {
+        if($month3 == "on") {
             $options += getDates(31*$_GET['day'], 3);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
             [
@@ -217,7 +233,7 @@ switch ($_GET['package']){
                 $month3
             ]);
         }
-        if($month4 === "on") {
+        if($month4 == "on") {
             $options += getDates(31*$_GET['day'], 4);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
             [
@@ -227,7 +243,7 @@ switch ($_GET['package']){
                 $month4
             ]);
         }
-        if($month5 === "on") {
+        if($month5 == "on") {
             $options += getDates(31*$_GET['day'], 5);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -237,7 +253,7 @@ switch ($_GET['package']){
                 $month5
             ]);
         }
-        if($month6 === "on") {
+        if($month6 == "on") {
             $options += getDates(31*$_GET['day'], 6);
              $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -247,7 +263,7 @@ switch ($_GET['package']){
                 $month6
             ]);
         }
-        if($month7 === "on") {
+        if($month7 == "on") {
             $options += getDates(31*$_GET['day'], 7);
              $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -257,8 +273,20 @@ switch ($_GET['package']){
                 $month7
             ]);
         }
+        $endService = date('Y-m-d', time() + $_GET['hour'] * 31 * 24 * 60 * 60);
 
-        $DbManager->getDb()->query("INSERT INTO orders(idOrders, type, price, status, idPerson, idService) VALUES(?, ?, ?, ?, ?, ?)",
+        $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
+            [
+                DbManager::v4(),
+                $uuid,
+                "timeVariable",
+                $endService
+            ]);
+
+        $price*=$options;
+        $finalPrice = $price + $price * 0.21;
+
+        $DbManager->query("INSERT INTO orders(idOrders, type, price, status, idPerson, idService) VALUES(?, ?, ?, ?, ?, ?)",
             [
             $uuid,
             substr($_GET['package'], 0, -3),
@@ -284,7 +312,7 @@ switch ($_GET['package']){
         $price = $_GET['hour'] * $chosenService[0]['price'] * 1 / 8;
         $options = 0;
 
-        if($year1 === "on") {
+        if($year1 == "on") {
             $options += getDates(31*12*$_GET['day'], 1);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -294,7 +322,7 @@ switch ($_GET['package']){
                 $year1
             ]);
         }
-        if($year2 === "on") {
+        if($year2 == "on") {
             $options += getDates(31*12*$_GET['day'], 2);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",[
                 DbManager::v4(),
@@ -303,7 +331,7 @@ switch ($_GET['package']){
                 $year2
             ]);
         }
-        if($year3 === "on") {
+        if($year3 == "on") {
             $options += getDates(31*12*$_GET['day'], 3);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -313,7 +341,7 @@ switch ($_GET['package']){
                 $year3
             ]);
         }
-        if($year4 === "on") {
+        if($year4 == "on") {
             $options += getDates(31*12*$_GET['day'], 4);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -323,7 +351,7 @@ switch ($_GET['package']){
                 $year4
             ]);
         }
-        if($year5 === "on") {
+        if($year5 == "on") {
             $options += getDates(31*12*$_GET['day'], 5);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -333,7 +361,7 @@ switch ($_GET['package']){
                 $year5
             ]);
         }
-        if($year6 === "on") {
+        if($year6 == "on") {
             $options += getDates(31*12*$_GET['day'], 6);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -343,7 +371,7 @@ switch ($_GET['package']){
                 $year6
             ]);
         }
-        if($year7 === "on") {
+        if($year7 == "on") {
             $options += getDates(31*12*$_GET['day'], 7);
             $q = $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
                 [
@@ -353,6 +381,15 @@ switch ($_GET['package']){
                 $year7
             ]);
         }
+        $endService = date('Y-m-d', time() + $_GET['hour'] * 31 * 12 * 24 * 60 * 60);
+
+        $DbManager->query("INSERT INTO orderoptions(idOrderOption, idOrders, typeOptions, options) VALUES (?, ?, ?, ?)",
+            [
+                DbManager::v4(),
+                $uuid,
+                "timeVariable",
+                $endService
+            ]);
 
         $price*=$options;
         $finalPrice = $price + $price * 0.21;
